backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
# 2.28+ accepts Supabase secret keys (sb_secret_...); 2.10 only accepted JWT service_role (eyJ...).
supabase==2.28.3
google-genai==1.70.0
anthropic==0.40.0
reportlab==4.4.10
python-jose[cryptography]==3.3.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic-settings==2.6.1
email-validator==2.2.0
json-repair==0.58.7
trafilatura==2.0.0
pypdf==5.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-cov==6.1.1
pytest-html==4.1.1
